% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IK_bandwidth}
\alias{IK_bandwidth}
\title{Compute MSE-optimal Imbens-Kalyanaraman bandwidth for a sharp RD.}
\usage{
IK_bandwidth(
  Y,
  X,
  threshold,
  kernel = c("triangular", "uniform", "epanechnikov")
)
}
\arguments{
\item{Y}{The outcomes.}

\item{X}{The running variable.}

\item{threshold}{The threshold.}

\item{kernel}{The kernel type used to construct weights within the bandwidth.}
}
\value{
A list containing the sample weights along with optimal bandwidth.
}
\description{
This convenience function computes weights using the Imbens-Kalyanaraman bandwidth procedure.
The code does exactly what the MATLAB code available on the author's website does.
}
\examples{
set.seed(42)
n = 1000; threshold = 0
X = runif(n, -1, 1)
W = as.numeric(X >= threshold)
Y = (1 + 2*W)*(1 + X^2) + 1 / (1 + exp(X)) + rnorm(n, sd = .5)
out = IK_bandwidth(Y, X, threshold)

}
\references{
Imbens, G., and Kalyanaraman, K. (2012).
 Optimal Bandwidth Choice for the Regression Discontinuity Estimator.
 The Review of Economic Studies, 79(3).
}
