% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dynMPE}
\alias{print.dynMPE}
\title{Print a dynMPE object}
\usage{
\method{print}{dynMPE}(x, digits = 4, ...)
}
\arguments{
\item{x}{dynMPE object}

\item{digits}{number of digits to print}

\item{...}{Additional arguments passed to print methods.}
}
\description{
Print a dynMPE object
}
