% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.dynMPE}
\alias{summary.dynMPE}
\title{dynMPE summary}
\usage{
\method{summary}{dynMPE}(object, ...)
}
\arguments{
\item{object}{dynMPE object}

\item{...}{Additional arguments (currently ignored).}
}
\description{
dynMPE summary
}
